final changes after review
