+++ BugTrackingSystem/settings.py
@@ -13,15 +13,10 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 SECRET_KEY = config('SECRET_KEY')
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = True
 
-# config('DEBUG', default=False, cast=bool)
-
-# handler404 = 'your_app.views.custom_404'
+DEBUG = config('DEBUG', cast=bool)
 
 ALLOWED_HOSTS = []
-
-# 'localhost', '127.0.0.1'
 
 # Application definition
