--- project/tests.py
@@ -1,1 +1,0 @@
-from django.test import TestCase # noqa
